+++ aiobungie/crate/__init__.py
@@ -107,12 +107,16 @@ __all__: tuple[str, ...] = (
     "ActivityValues",
     "PostActivityPlayer",
     "PostActivityTeam",
+    "ExtendedValues",
+    "ExtendedWeaponValues",
 )
 
 from .activity import Activity
 from .activity import ActivityValues
 from .activity import AvailableActivity
 from .activity import Challenges
+from .activity import ExtendedValues
+from .activity import ExtendedWeaponValues
 from .activity import GuidedGame
 from .activity import Location
 from .activity import Matchmaking