+++ CHANGELOG.md
@@ -12,6 +12,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - `Character.fetch_activities` method.
 - Finished implementing post activities methods.
 - `is_flawless`, `is_solo` and `is_solo_flawless` useful properties to for both `Activity` and `PostActivity`.
+- Post activity extended values and player weapons values.
+- `rest._handle_error` method moved to `error.py` and renamed to `raise_error`
 
 ### Changed
 - Exceptions now has fields and improved.